Component class WIP
