+++ www/examples/templates/index.html
@@ -43,7 +43,8 @@
           background-color: DarkSlateGray;
         }
       </style>
-      <p class="w3-padding-small">Template style tags affect the rest of the page</p>
+      <p class="w3-padding-small">
+        Template style tags affect the rest of the page</p>
     </div>
   </template>
